Make countAll return a long

+++ domain/src/commonMain/kotlin/org/sourcegrade/kontour/Repository.kt
@@ -24,7 +24,7 @@ import kotlin.reflect.KClass
 interface Repository<E : DomainEntity, C : Creates<E>> {
     suspend fun create(item: C): E
     suspend fun exists(id: UUID): Boolean
-    suspend fun countAll(): Int
+    suspend fun countAll(): Long
     suspend fun <D : Dto<E>> findDtoById(id: UUID, dtoType: KClass<D>): D?
     suspend fun deleteById(id: UUID): Boolean
 }
